+++ smartdada2/testing/GetTrimParameters.py
